+++ wxc_sdk/telephony/callqueue/__init__.py
@@ -415,8 +415,7 @@ class AvailableAgent(ApiModel):
     #: Person has the CX Essentials license.
     has_cx_essentials: Optional[bool] = None
     #: List of phone numbers of a person, workspace or virtual line.
-    #: TODO name change, WXCAPIBULK-697
-    numbers: Optional[list[UserNumber]] = Field(default=None, alias='phoneNumbers')
+    phone_numbers: Optional[list[UserNumber]] = None
 
 
 @dataclass(init=False, repr=False)